fix(services): split FontLoader so expo-font never enters the web graph
`FontLoader.tsx` did an unconditional top-level `import * as Font from
'expo-font'` with no platform fork, so the published `lib/module/.../
FontLoader.js` dragged expo-font into the browser bundle. `expo-font/build/
serverContext.web.js` opens with `import { AsyncLocalStorage } from
'node:async_hooks'`, which every browser bundler externalizes — touching the
binding throws at module-evaluation time and React never mounts:

  Module "node:async_hooks" has been externalized for browser compatibility.
  Cannot access "node:async_hooks.AsyncLocalStorage" in client code.

Same class of bug as the vector-icons peer fix, and expo-font's own
`getStorage()` already throws when `window` is defined — that module is
server-only by design and should never be reachable on web.

Split along the established `.native`/clean-default convention already used by
`webauthn/passkeyClient`:
- `FontLoader.native.tsx` keeps the real expo-font `loadAsync` implementation.
- `FontLoader.tsx` is the clean default (web + SSR): `FontLoader` renders its
  children, `setupFonts` is a no-op returning true. It imports nothing from
  expo-font, so the web bundle can no longer reach node:async_hooks.

Nothing is lost on web: web fonts are Bloom's job — `applyFontFaces()` in
@oxyhq/bloom injects the Inter/BlomusModernus/Geist Mono @font-face rules as
base64 data URLs. The old web branch here (dynamic @font-face injection from
`require()`d .ttf) was already dead: its `require` calls threw straight into
the surrounding catch under any web bundler.

expo-font stays a non-optional peer (it is genuinely required on native and
already declared as such); this changes only which platforms pull it.
